+++ core/src/main/java/com/sonicgdx/sonicswirl/Player.java
@@ -24,8 +24,8 @@ public class Player extends Entity {
     public void move(float delta)
     {
 
-        SensorReturn leftSensorTile = downSensorCheck((int) (xPos), (int) yPos);
-        SensorReturn rightSensorTile = downSensorCheck((int) (xPos + sprite.getWidth()), (int) yPos);
+        SensorReturn leftSensorTile = downSensorCheck(xPos, yPos);
+        SensorReturn rightSensorTile = downSensorCheck(xPos + sprite.getWidth(), yPos);
         Gdx.app.debug("Right Ground Sensor distance", String.valueOf(rightSensorTile.returnDistance));
 
         //TODO Would be better to implement an InputProcessor. This makes more sense as an interrupt rather than constant polling.
